pyproject.toml
[tool.poetry]
name = "genai-python"
version = "0.1.0"
description = ""
authors = ["devnaveedsarwar <learnwithnaveedsarwar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.2.29"
langchain-community = "^0.2.11"
langchain-openai = "^0.1.21"
python-dotenv = "^1.0.1"
langchain-google-genai = "^1.0.9"
langchain-google-vertexai = "^1.0.8"
huggingface-hub = "^0.24.5"
transformers = "^4.44.0"
langchain-huggingface = "^0.0.3"
bs4 = "^0.0.2"
langchain-chroma = "^0.1.3"
faiss-cpu = "^1.8.0.post1"
langchainhub = "^0.1.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
